refactor: add project prefix to all functions
Added lngrd_ prefix to all static functions. Turns out that static
functions in a header can conflict with outside code within the same
translation unit. To avoid potential naming conflicts with other
functions in a program a prefix is added as a form of namespacing to
reduce the risk of overlap.